+++ Source/LogBridge.EnterpriseLibrary.Tests.Unit/LogDataVerifier.cs
@@ -69,6 +69,11 @@ namespace SoftwarePassion.LogBridge.EnterpriseLibrary.Tests.Unit
         private void CompareProperties(IDictionary<string, object> expected, IDictionary<string, object> actual)
         {
             // It is okay for the actual to have more, but it must have all from expected.
+            var expectedAsStrings = expected
+                .ToDictionary(item => item.Key, item => item.Value?.ToString());
+            var actualAsStrings = actual
+                .ToDictionary(item => item.Key, item => item.Value?.ToString());
+
             var expectedKeys = expected.Keys;
             var actualKeys = actual.Keys;
 
@@ -77,7 +82,7 @@ namespace SoftwarePassion.LogBridge.EnterpriseLibrary.Tests.Unit
                 .ToList();
 
             var nonMatchingKeys = expectedKeys
-                .Where(key => !Equals(expected[key], actual[key]))
+                .Where(key => !Equals(expectedAsStrings[key], actualAsStrings[key]))
                 .ToList();
 
             missingKeys.Count().Should().Be(0, because: "Missing properties: " + string.Join(", ", missingKeys));