+++ Source/LogBridge.Log4Net.Tests.Unit/LogDataVerifier.cs
@@ -23,7 +23,7 @@ namespace SoftwarePassion.LogBridge.Log4Net.Tests.Unit
             
             var actual = appender.GetEvents().First().GetLoggingEventData();
 
-            actual.TimeStamp.Should().Be(expected.TimeStamp, because: "Timestamp should match.");
+            actual.TimeStampUtc.Should().Be(expected.TimeStamp, because: "Timestamp should match.");
             actual.Properties[LogConstants.EventIdKey].Should().Be(expected.EventId);
 
             if (expected.CorrelationId.IsSome)
@@ -74,6 +74,9 @@ namespace SoftwarePassion.LogBridge.Log4Net.Tests.Unit
         private void CompareProperties(Dictionary<string, object> expected, PropertiesDictionary actual)
         {
             // It is okay for the actual to have more, but it must have all from expected.
+            var expectedAsStrings = expected
+                .ToDictionary(item => item.Key, item => item.Value?.ToString());
+
             var expectedKeys = expected.Keys;
             var actualKeys = actual.GetKeys();
 
@@ -82,7 +85,7 @@ namespace SoftwarePassion.LogBridge.Log4Net.Tests.Unit
                 .ToList();
 
             var nonMatchingKeys = expectedKeys
-                .Where(key => !Equals(expected[key], actual[key]))
+                .Where(key => !Equals(expectedAsStrings[key], actual[key].ToString()))
                 .ToList();
 
             missingKeys.Count().Should().Be(0, because: "Missing properties: " + string.Join(", ", missingKeys));