+++ Source/LogBridge.UmbracoLog4Net.Tests.Unit/LogDataVerifier.cs
@@ -82,7 +82,7 @@ namespace SoftwarePassion.LogBridge.UmbracoLog4Net.Tests.Unit
                 .ToList();
 
             var nonMatchingKeys = expectedKeys
-                .Where(key => !Equals(expected[key], actual[key]))
+                .Where(key => !Equals(expected[key].ToString(), actual[key].ToString()))
                 .ToList();
 
             missingKeys.Count().Should().Be(0, because: "Missing properties: " + string.Join(", ", missingKeys));